Update index.html
the logo is not only visually appealing but also strategically aligned with the brand’s goals. As a contributor, collaboration is key.  My role is to contribute my creative ideas.
Reasons for Changing a Logo:
Rebranding
Modernization
Simplification
I am passionate about the project and Thankful to you for giving a chance to contribute.